+++ app/src/main/java/com/madness/hobbymatcher/fragment/AddActivityFragment.kt
@@ -1,4 +1,4 @@
-package com.madness.hobbymatcher.AddActivity
+package com.madness.hobbymatcher.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,9 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
-import androidx.core.view.get
 import androidx.navigation.findNavController
-import androidx.navigation.fragment.navArgs
 import com.madness.hobbymatcher.HobbyMatcherApplication
 import com.madness.hobbymatcher.R
 import com.madness.hobbymatcher.networking.ActivityService
