+++ app/src/main/res/layout/fragment_add_activity.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddActivity.AddActivityFragment">
+    tools:context=".fragment.AddActivityFragment">
 
     <ScrollView
         android:layout_width="match_parent"
